\name{cosmoFns-package}
\alias{cosmoFns-package}
\alias{cosmoFns}
\docType{package}
\title{
Cosmology functions
}
\description{
Package contains functions for computation of distances and luminosities
in a flat cosmology.
}
\details{
\tabular{ll}{
Package: \tab cosmoFns\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2012-09-16\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
A. Harris

Maintainer: <harris@astro.umd.edu>
}
\references{
"Distance Measures in Cosmology," D.W. Hogg (2000),
arXiv:astro-ph/9905116;
"Warm Molecular Gas in the Pirmeval Galaxy 10214+4724", P.M. Solomon,
D. Downes, and S.J.E. Radford (1992), Ap.J. 398, L29;
"First-year WMAP observations...", Spergel et al., ApJS 148:175 (2003).
}
\examples{
D.L(z=2.3)
}
